test(web): add working middleware tests
