+++ projects/rawkode-academy/web/src/middleware/index.ts
@@ -1,0 +1,6 @@
+import { sequence } from "astro:middleware";
+import { authMiddleware } from "./auth";
+
+export const onRequest = sequence(
+	authMiddleware
+);
